+++ src/slices/recipes.js
@@ -7,10 +7,10 @@ export const fetchRandomRecipe = createAsyncThunk(
     'recipes/fetchRandomRecipe',
     async (number) => {
       const response = await recipeApi.generateRandomRecipe(number)
-      // console.log('Recipe Response', response)
+      console.log('Recipe Response', response)
       return response
     }
-  )
+)
 
 const initialState={
     recipeData:null,