+++ src/views/homePage/Result.jsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, {useState} from 'react'
 import SlideShow from '../../components/slideshow/SlideShow';
 
 import { 
@@ -9,6 +9,7 @@ import {
     Typography,
     useTheme,
     useMediaQuery,
+    LinearProgress,
 } from '@material-ui/core';
 
 import FoodGallery from '../../components/RecipeDisplay/FoodGallery'
@@ -16,9 +17,11 @@ import HeaderDivider from '../../components/HeaderDivider'
 
 import {
     reduxDispatch,
+    reduxSelector,
 } from '../../store/index'
 
 import { fetchRandomRecipe } from '../../slices/recipes'
+import ErrorBar from '../errors/ErrorBar'
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -30,14 +33,43 @@ export default function Result() {
     const dispatch=reduxDispatch()
     const theme = useTheme()
     const mobileDevice = useMediaQuery(theme.breakpoints.down('sm'))
+    const isLoading =reduxSelector(store=>store.recipes.loading)
+
+    const [errorState, setErrorState] =useState({
+        open:false,
+        msg:'Something went Wrong!'
+    });
+
+    const handleClose = (event, reason) => {
+        if (reason === 'clickaway') {
+          return
+        }
+        setErrorState({
+            ...errorState,
+            open:false
+        })
+    };
+
     return (
         <React.Fragment>
             <Container 
                 maxWidth="lg"
                 className={classes.root}
             >
+                {errorState.open&& 
+                <ErrorBar 
+                    errorState={errorState}
+                    handleClose={handleClose}
+                />}
+
                 <Box mt={5}>
-                    <SlideShow/>
+                    <Typography align='center' gutterBottom={true}>
+                            {mobileDevice ?<h4>Featured articles:</h4>:<h3>Featured articles:</h3>}
+                    </Typography>
+                    <HeaderDivider/>
+                    <Box m={2} align='center'>
+                        <SlideShow/>
+                    </Box>
                 </Box>
 
                 <Box mt={5}>
@@ -48,8 +80,22 @@ export default function Result() {
                     <HeaderDivider/>
                     <Box m={2} align='center'>
                     <Button
-                        onClick={()=>{
-                            dispatch(fetchRandomRecipe(4))
+                        onClick={async()=>{
+                            try{
+                                const result = await dispatch(fetchRandomRecipe(4))
+                                if(result.error){
+                                    // console.log(result.error)
+                                    const errorType =result.error.message.split("")
+                                    const msg =errorType[errorType.length-1]==='404'?'Error Code 404 bad request':'Reach max amount of request for current Spoonacular API Plan'
+
+                                    setErrorState({
+                                        msg:msg,
+                                        open:true
+                                    })
+                                }
+                            }catch(e){
+                                // console.log(e)
+                            }
                         }}
                         color='secondary' 
                         variant="outlined" 
@@ -57,7 +103,11 @@ export default function Result() {
                         Generate Random Recipes
                     </Button>
                     </Box>
-                    <FoodGallery />
+
+                    {isLoading
+                    ?<LinearProgress color="secondary" />
+                    :<FoodGallery />
+                    }
                 </Box>
 
             </Container>